+++ CHANGELOG.md
@@ -39,7 +39,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- Nothing.
+- [#40](https://github.com/zendframework/zend-cache/pull/40)
+  Fixed TTL support of `Redis::addItem`
 
 ## 2.5.3 - 2015-09-15